Add a counter for deleted documents to FragmentedTextSegment (#26366)
We need this counter for three things:

1. To avoid reuploading unmodified segments during a flush
2. To detect collisions between compaction and flushing
3. To accurately log the number of documents in each segment

GitOrigin-RevId: 9c5fb6224a0a260e2677507b5fd4917e75e9e670

+++ crates/database/src/index_workers/index_meta.rs
@@ -50,7 +50,7 @@ pub trait PreviousSegmentsType: Send {
 pub trait SegmentType<T: SearchIndex> {
     fn id(&self) -> &str;
 
-    fn num_deleted(&self) -> u32;
+    fn num_deleted(&self) -> u64;
 
     fn statistics(&self) -> anyhow::Result<T::Statistics>;
 }
